[+] introduce new type `metrics.MetricIntervals` (#1312)
We want to change it to `map[string]int` instead of `map[string]float64`
and use the type name everywhere in code base to make the purpose of
vars obvious

+++ internal/reaper/logparser.go
@@ -43,7 +43,7 @@ type LogParser struct {
 	ctx              context.Context
 	LogsMatchRegex   *regexp.Regexp
 	SourceConn       *sources.SourceConn
-	Interval         float64
+	Interval         int
 	StoreCh          chan<- metrics.MeasurementEnvelope
 	eventCounts      map[string]int64 // for the specific DB. [WARNING: 34, ERROR: 10, ...], zeroed on storage send
 	eventCountsTotal map[string]int64 // for the whole instance
